% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsapWeights.R
\name{deltaMatrix_from_BRLchain}
\alias{deltaMatrix_from_BRLchain}
\title{Compute delta matrix from BRL chain}
\usage{
deltaMatrix_from_BRLchain(Zchain, n1, n2, k, beta = 1)
}
\description{
Compute delta matrix from BRL chain
}
