% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsapWeights.R
\name{deltaMatrix_from_list}
\alias{deltaMatrix_from_list}
\title{Compute delta matrix from list of coreference matrices}
\usage{
deltaMatrix_from_list(matrixChain, n1, n2, k, beta = 1)
}
\description{
Compute delta matrix from list of coreference matrices
}
\note{
List of coreference matrices can be computed from a chain of BRL vectors as:
\code{
matrices = lapply(1:dim(chains)[[2]], function(i){
  Z = chains[, i]
  return(weightsMatrix_from_BRLVector(Z, n1, n2))
})
}
}
