% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStructures.R
\name{isBipartiteLinkageVector}
\alias{isBipartiteLinkageVector}
\title{Check that object is a bipartite linkage vector.}
\usage{
isBipartiteLinkageVector(Z, n1, n2)
}
\arguments{
\item{Z}{object check.}

\item{n1}{size of the first list of records.}

\item{n2}{size of the second list of records.}
}
\value{
TRUE if \code{Z} is a bipartite linkage vector, FALSE otherwise.
}
\description{
A bipartite linkage vector \code{Z} represents a bipartite linkage between two lists of records of sizes \code{n1} and \code{n2}. It is a vector of length \code{n2} defined as follows. For a given index \code{j}, where \verb{1 <= j <= n2}, \code{Z[j] = i} if record \code{j} is linked to record \code{i}. If record \code{j} is not linked to any other record, then \code{Z_j = n_1 + j}.
}
\details{
Therefore, a valid bipartite linkage vector for lists of sizes \code{n1} and \code{n2} should satisfy the following requirements:
\enumerate{
\item It is an integer vector of length \code{n2}.
\item Values \code{Z[j]} satisfy \verb{1 <= Z[j] <= 2} or \code{Z[j] == n2 + j}.
\item Since \code{Z} should represent a bipartite linkage, \code{Z} should not have any duplicate values.
}
}
\examples{
Z <- c(1, 2, 4)
n1 <- 4
n2 <- 3
isBipartiteLinkageVector(Z, n1, n2)
# Returns: TRUE

}
