% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStructures.R
\name{isLinkageBigraph}
\alias{isLinkageBigraph}
\title{Check that graph is a bipartite linkage graph.}
\usage{
isLinkageBigraph(graph, n1, n2)
}
\arguments{
\item{graph}{object check.}

\item{n1}{size of the first list of records.}

\item{n2}{size of the second list of records.}
}
\value{
TRUE if \code{graph} is a bipartite linkage graph, FALSE otherwise.
}
\description{
A bipartite linkage graph is a bipartite graph with degrees at most 1.
}
\examples{
graph <- igraph::graph.bipartite(types = c(rep(TRUE, 3), rep(FALSE, 2)), edges = c(1, 4), directed = FALSE)
n1 <- 3
n2 <- 2
isLinkageBigraph(graph, n1, n2)
# Returns: TRUE

graph[]

# 5 x 5 sparse Matrix of class 'dgCMatrix'
#
# [1,] . . . 1 .
# [2,] . . . . .
# [3,] . . . . .
# [4,] 1 . . . .
# [5,] . . . . .

}
