% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStructures.R
\name{isWeightedBigraph}
\alias{isWeightedBigraph}
\title{Check that graph is a bipartite weighted graph.}
\usage{
isWeightedBigraph(graph, n1, n2)
}
\arguments{
\item{graph}{object check.}

\item{n1}{size of the first list of records.}

\item{n2}{size of the second list of records.}
}
\value{
TRUE if \code{graph} is a bipartite weighted graph, FALSE otherwise.
}
\description{
A bipartite weighted graph is a bipartite weighted graph between lists of sizes \code{n1} and \code{n2}.
}
\examples{
graph <- igraph::graph.bipartite(types = c(rep(TRUE, 3), rep(FALSE, 2)), edges = c(1, 4, 3, 5), directed = FALSE)
igraph::E(graph)$weight <- c(1.5, 0.5)
n1 <- 3
n2 <- 2
isWeightedBigraph(graph, n1, n2)
# Returns: TRUE

graph[]
# 5 x 5 sparse Matrix of class 'dgCMatrix'
#
# [1,] .   . .   1.5 .
# [2,] .   . .   .   .
# [3,] .   . .   .   0.5
# [4,] 1.5 . .   .   .
# [5,] .   . 0.5 .   .

}
