% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStructures.R
\name{isWeightsMatrix}
\alias{isWeightsMatrix}
\title{Check that matrix represents weights between two record lists.}
\usage{
isWeightsMatrix(matrix, n1, n2)
}
\arguments{
\item{matrix}{object check. Can be a sparse matrix.}

\item{n1}{size of the first list of records.}

\item{n2}{size of the second list of records.}
}
\value{
TRUE if \code{matrix} is a bipartite weights matrix, FALSE otherwise.
}
\description{
A weights matrix is a matrix with rows corresponding to the \code{n1} elements of a first list, and columns corresponding to the \code{n2} elements to the second list.
}
\examples{
matrix <- Matrix::Matrix(c(0, 1, 0, 0, 1, 0), nrow = 3, byrow = TRUE)
matrix
# 3 x 2 sparse Matrix of class 'dgCMatrix'
#
# [1,] . 1
# [2,] . .
# [3,] 1 .

n1 <- 3
n2 <- 2
isWeightsMatrix(matrix, n1, n2)
# Returns: TRUE

}
