% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsap.R
\name{lsap2}
\alias{lsap2}
\title{Run LSAP (conditional on number of links) from a weights matrix.}
\usage{
lsap2(weightsMatrix, k = NULL, C = 1e+06)
}
\description{
Run LSAP (conditional on number of links) from a weights matrix.
}
